updated arduino for maxvalue and action set to integer

+++ scripts/mainclasses.py
@@ -561,13 +561,13 @@ class Controls(object):
             if joybuttons:
                 if joybuttons[7] == 0 and joybuttons[8] == 0:
                     if joyaxes[2] < -0.1:
-                        base.action = 300 - (joyaxes[2] * 99)
+                        base.action = int(300 - (joyaxes[2] * 99))
                     elif joyaxes[2] >  0.1:
-                        base.action = 400 + (joyaxes[2] * 99)
+                        base.action = int(400 + (joyaxes[2] * 99))
                     elif joyaxes[3] >  0.1:
-                        base.action = 100 + (joyaxes[3] * 99)
+                        base.action = int(100 + (joyaxes[3] * 99))
                     elif joyaxes[3] <  -0.1:
-                        base.action = 200 - (joyaxes[3] * 99)
+                        base.action = int(200 - (joyaxes[3] * 99))
                     else:
                         base.action = 0
                     print(base.action)
